Restructured API clients into Infrastructure entirely.
