--- Skylight/Skylight-Application/Interfaces/Infrastructure/Clients/NationalWeatherService/AlertCertainty.cs
@@ -1,13 +1,0 @@
-﻿namespace Skylight.Application.Interfaces.Infrastructure.Clients.NationalWeatherService;
-
-/// <summary>
-/// Alert certainty.
-/// </summary>
-public enum AlertCertainty
-{
-    Observed,
-    Likely,
-    Possible,
-    Unlikely,
-    Unknown
-}
