--- Skylight/Skylight-Application/Interfaces/Infrastructure/Clients/NationalWeatherService/AlertMessageType.cs
@@ -1,11 +1,0 @@
-﻿namespace Skylight.Application.Interfaces.Infrastructure.Clients.NationalWeatherService;
-
-/// <summary>
-/// Alert message type.
-/// </summary>
-public enum AlertMessageType
-{
-    Alert,
-    Update,
-    Cancel
-}
