--- Skylight/Skylight-Application/Interfaces/Infrastructure/Clients/NationalWeatherService/AlertSeverity.cs
@@ -1,13 +1,0 @@
-﻿namespace Skylight.Application.Interfaces.Infrastructure.Clients.NationalWeatherService;
-
-/// <summary>
-/// Alert severity.
-/// </summary>
-public enum AlertSeverity
-{
-    Extreme,
-    Severe,
-    Moderate,
-    Minor,
-    Unknown
-}
