--- Skylight/Skylight-Application/Interfaces/Infrastructure/Clients/NationalWeatherService/AlertStatus.cs
@@ -1,13 +1,0 @@
-﻿namespace Skylight.Application.Interfaces.Infrastructure.Clients.NationalWeatherService;
-
-/// <summary>
-/// Alert status.
-/// </summary>
-public enum AlertStatus
-{
-    Actual,
-    Exercise,
-    System,
-    Test,
-    Draft,
-}
