--- Skylight/Skylight-Application/Interfaces/Infrastructure/Clients/NationalWeatherService/AlertUrgency.cs
@@ -1,13 +1,0 @@
-﻿namespace Skylight.Application.Interfaces.Infrastructure.Clients.NationalWeatherService;
-
-/// <summary>
-/// Alert urgency.
-/// </summary>
-public enum AlertUrgency
-{
-    Immediate,
-    Expected,
-    Future,
-    Past,
-    Unknown
-}
