--- Skylight/Skylight-Application/Interfaces/Infrastructure/Clients/NationalWeatherService/MarineAreaCode.cs
@@ -1,82 +1,0 @@
-﻿namespace Skylight.Application.Interfaces.Infrastructure.Clients.NationalWeatherService;
-
-/// <summary>
-/// Marine area codes as defined in NWS Directive 10-302.
-/// </summary>
-public enum MarineAreaCode
-{
-    /// <summary>
-    /// Western North Atlantic Ocean and along U.S. East Coast south of Currituck Beach Light NC following the coastline into Gulf of Mexico to Ocean Reef FL including the Caribbean.
-    /// </summary>
-    AM,
-
-    /// <summary>
-    /// Western North Atlantic Ocean and along U.S. East Coast from Canadian border south to Currituck Beach Light NC.
-    /// </summary>
-    AN,
-
-    /// <summary>
-    /// Gulf of Mexico and along the U.S. Gulf Coast from the Mexican border to Ocean Reef FL.
-    /// </summary>
-    GM,
-
-    /// <summary>
-    /// Lake St. Clair.
-    /// </summary>
-    LC,
-
-    /// <summary>
-    /// Lake Erie.
-    /// </summary>
-    LE,
-
-    /// <summary>
-    /// Lake Huron.
-    /// </summary>
-    LH,
-
-    /// <summary>
-    /// Lake Michigan.
-    /// </summary>
-    LM,
-
-    /// <summary>
-    /// Lake Ontario.
-    /// </summary>
-    LO,
-
-    /// <summary>
-    /// Lake Superior.
-    /// </summary>
-    LS,
-
-    /// <summary>
-    /// Central Pacific Ocean including Hawaiian waters.
-    /// </summary>
-    PH,
-
-    /// <summary>
-    /// North Pacific Ocean near Alaska and along Alaska coastline including the Bering Sea and the Gulf of Alaska.
-    /// </summary>
-    PK,
-
-    /// <summary>
-    /// Western Pacific Ocean including Mariana Island waters.
-    /// </summary>
-    PM,
-
-    /// <summary>
-    /// South Central Pacific Ocean including American Samoa waters.
-    /// </summary>
-    PS,
-
-    /// <summary>
-    /// Eastern North Pacific Ocean and along U.S. West Coast from Canadian border to Mexican border.
-    /// </summary>
-    PZ,
-
-    /// <summary>
-    /// St. Lawrence River above St. Regis.
-    /// </summary>
-    SL
-}
