--- Skylight/Skylight-Application/Interfaces/Infrastructure/Clients/NationalWeatherService/RegionType.cs
@@ -1,10 +1,0 @@
-﻿namespace Skylight.Application.Interfaces.Infrastructure.Clients.NationalWeatherService;
-
-/// <summary>
-/// Region type.
-/// </summary>
-public enum RegionType
-{
-    Land,
-    Marine
-}
