+++ Skylight/Skylight-Architecture-Test/Application/StructureTests.cs
@@ -21,22 +21,4 @@ public class StructureTests(ITestOutputHelper outputHelper)
 
         Assert.True(result.IsSuccessful);
     }
-
-    [Fact]
-    public void ClientModels_Should_BeSealed()
-    {
-        TestResult result = Types
-            .InAssembly(Assemblies.Application)
-            .That()
-            .ResideInNamespaceContaining(nameof(Skylight.Application.Interfaces.Infrastructure.Clients))
-            .And()
-            .AreNotAbstract()
-            .Should()
-            .BeSealed()
-            .GetResult();
-
-        TestOutputHelpers.PrintFailingTypes(outputHelper, result);
-
-        Assert.True(result.IsSuccessful);
-    }
 }
