+++ Skylight/Skylight-Architecture-Test/Assemblies.cs
@@ -10,4 +10,6 @@ public static class Assemblies
     public static readonly Assembly Domain = typeof(Skylight.Domain.Entities.BaseEntity).Assembly;
 
     public static readonly Assembly Application = typeof(Skylight.Application.DependencyInjection).Assembly;
+
+    public static readonly Assembly Infrastructure = typeof(Skylight.Infrastructure.DependencyInjection).Assembly;
 }
