+++ Skylight/Skylight-Infrastructure/Clients/IClientRequest.cs
@@ -1,0 +1,6 @@
+﻿namespace Skylight.Infrastructure.Clients;
+
+/// <summary>
+/// Denotes a request to a third-party API.
+/// </summary>
+public interface IClientRequest { }
