+++ Skylight/Skylight-Infrastructure/Clients/IClientResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace Skylight.Infrastructure.Clients;
+
+/// <summary>
+/// Denotes a response from a third-party API.
+/// </summary>
+public interface IClientResponse { }
