+++ Skylight/Skylight-Infrastructure/Clients/NationalWeatherService/Models/AlertCertainty.cs
@@ -1,0 +1,13 @@
+﻿namespace Skylight.Infrastructure.Clients.NationalWeatherService.Models;
+
+/// <summary>
+/// Alert certainty.
+/// </summary>
+public enum AlertCertainty
+{
+	Observed,
+	Likely,
+	Possible,
+	Unlikely,
+	Unknown
+}
