+++ Skylight/Skylight-Infrastructure/Clients/NationalWeatherService/Models/AlertCollection.cs
@@ -1,0 +1,9 @@
+﻿namespace Skylight.Infrastructure.Clients.NationalWeatherService.Models;
+
+/// <summary>
+/// A group of <see cref="Alert"/> instances.
+/// </summary>
+public sealed record AlertCollection(
+	string Title,
+	DateTimeOffset Updated,
+	IReadOnlyList<Alert> Alerts);
