+++ Skylight/Skylight-Infrastructure/Clients/NationalWeatherService/Models/AlertMessageType.cs
@@ -1,0 +1,11 @@
+﻿namespace Skylight.Infrastructure.Clients.NationalWeatherService.Models;
+
+/// <summary>
+/// Alert message type.
+/// </summary>
+public enum AlertMessageType
+{
+	Alert,
+	Update,
+	Cancel
+}
