+++ Skylight/Skylight-Infrastructure/Clients/NationalWeatherService/Models/AlertResponse.cs
@@ -1,17 +1,17 @@
-﻿namespace Skylight.Application.Interfaces.Infrastructure.Clients.NationalWeatherService;
+﻿namespace Skylight.Infrastructure.Clients.NationalWeatherService.Models;
 
 /// <summary>
 /// The code denoting the type of action recommended for the target audience. This corresponds to <c>responseType</c> in the CAP specification. 
 /// </summary>
 public enum AlertResponse
 {
-    None,
-    AllClear,
-    Assess,
-    Monitor,
-    Avoid,
-    Execute,
-    Prepare,
-    Evacuate,
-    Shelter
+	None,
+	AllClear,
+	Assess,
+	Monitor,
+	Avoid,
+	Execute,
+	Prepare,
+	Evacuate,
+	Shelter
 }
