+++ Skylight/Skylight-Infrastructure/Clients/NationalWeatherService/Models/AlertSeverity.cs
@@ -1,0 +1,13 @@
+﻿namespace Skylight.Infrastructure.Clients.NationalWeatherService.Models;
+
+/// <summary>
+/// Alert severity.
+/// </summary>
+public enum AlertSeverity
+{
+	Extreme,
+	Severe,
+	Moderate,
+	Minor,
+	Unknown
+}
