+++ Skylight/Skylight-Infrastructure/Clients/NationalWeatherService/Models/AlertStatus.cs
@@ -1,0 +1,13 @@
+﻿namespace Skylight.Infrastructure.Clients.NationalWeatherService.Models;
+
+/// <summary>
+/// Alert status.
+/// </summary>
+public enum AlertStatus
+{
+	Actual,
+	Exercise,
+	System,
+	Test,
+	Draft,
+}
