+++ Skylight/Skylight-Infrastructure/Clients/NationalWeatherService/Models/AlertUrgency.cs
@@ -1,0 +1,13 @@
+﻿namespace Skylight.Infrastructure.Clients.NationalWeatherService.Models;
+
+/// <summary>
+/// Alert urgency.
+/// </summary>
+public enum AlertUrgency
+{
+	Immediate,
+	Expected,
+	Future,
+	Past,
+	Unknown
+}
