+++ Skylight/Skylight-Infrastructure/Clients/NationalWeatherService/Models/MarineRegionCode.cs
@@ -1,0 +1,37 @@
+﻿namespace Skylight.Infrastructure.Clients.NationalWeatherService.Models;
+
+/// <summary>
+/// Combined groups of <see cref="MarineAreaCode"/>.
+/// </summary>
+public enum MarineRegionCode
+{
+	/// <summary>
+	/// Alaska Waters (PK).
+	/// </summary>
+	AL,
+
+	/// <summary>
+	/// Atlantic Ocean (AM, AN).
+	/// </summary>
+	AT,
+
+	/// <summary>
+	/// Great Lakes (LC, LE, LH, LM, LO, LS, SL).
+	/// </summary>
+	GL,
+
+	/// <summary>
+	/// Gulf of Mexico (GM).
+	/// </summary>
+	GM,
+
+	/// <summary>
+	/// Eastern Pacific Ocean and U.S. West Coast (PZ).
+	/// </summary>
+	PA,
+
+	/// <summary>
+	/// Central and Western Pacific (PH, PM, PS).
+	/// </summary>
+	PI,
+}
