+++ Skylight/Skylight-Infrastructure/Clients/NationalWeatherService/Models/RegionType.cs
@@ -1,0 +1,10 @@
+﻿namespace Skylight.Infrastructure.Clients.NationalWeatherService.Models;
+
+/// <summary>
+/// Region type.
+/// </summary>
+public enum RegionType
+{
+	Land,
+	Marine
+}
