+++ Skylight/Skylight-Infrastructure/Clients/NationalWeatherService/Models/StateTerritoryCode.cs
@@ -1,0 +1,67 @@
+﻿namespace Skylight.Infrastructure.Clients.NationalWeatherService.Models;
+
+/// <summary>
+/// State territory codes.
+/// </summary>
+public enum StateTerritoryCode
+{
+	AL,
+	AK,
+	AS,
+	AR,
+	AZ,
+	CA,
+	CO,
+	CT,
+	DE,
+	DC,
+	FL,
+	GA,
+	GU,
+	HI,
+	ID,
+	IL,
+	IN,
+	IA,
+	KS,
+	KY,
+	LA,
+	ME,
+	MD,
+	MA,
+	MI,
+	MN,
+	MS,
+	MO,
+	MT,
+	NE,
+	NV,
+	NH,
+	NJ,
+	NM,
+	NY,
+	NC,
+	ND,
+	OH,
+	OK,
+	OR,
+	PA,
+	PR,
+	RI,
+	SC,
+	SD,
+	TN,
+	TX,
+	UT,
+	VT,
+	VI,
+	VA,
+	WA,
+	WV,
+	WI,
+	WY,
+	MP,
+	PW,
+	FM,
+	MH
+}
